+++ fleche/config.ml
@@ -16,6 +16,11 @@ type t =
         (** Show `msg_info` messages in diagnostics *)
   ; show_notices_as_diagnostics : bool [@default false]
         (** Show `msg_notice` messages in diagnostics *)
+  ; admit_on_bad_qed : bool [@default true]
+        (** [admit_on_bad_qed] There are two possible error recovery strategies
+            when a [Qed] fails: one is not to add the constant to the state, the
+            other one is admit it. We find the second behavior more useful, but
+            YMMV. *)
   }
 
 let default =
@@ -27,6 +32,7 @@ let default =
   ; goal_after_tactic = false
   ; show_coq_info_messages = false
   ; show_notices_as_diagnostics = false
+  ; admit_on_bad_qed = true
   }
 
 let v = ref default